pyproject.toml
[project]
name = "rag-real-proejct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-mistralai>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.9.1",
    "python-dotenv>=1.0.1",
    "python-magic>=0.4.27 ; sys_platform != 'win32'",
    "python-magic-bin>=0.4.14 ; sys_platform == 'win32'",
    "rapidocr-onnxruntime==1.2.3",
    "sentence-transformers==5.3.0",
    "streamlit>=1.35.0",
    "unstructured[all-docs]>=0.21.5",
]
